메소드 오버라이딩 예제 추가

+++ src/com/gyojincompany/ch07/part01/Shape.java
@@ -5,8 +5,8 @@ public class Shape { //부모클래스
 	int height;
 	String color; //도형면의 색깔
 	
-	public int area() { //도형의 면적 계산 메소드
-		int area = width * height;
+	public double area() { //도형의 면적 계산 메소드
+		double area = width * height;
 		return area;
 	}
 }
